% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_CDMConnector.R
\name{cdm_getLatestDataFreezeAndVersion}
\alias{cdm_getLatestDataFreezeAndVersion}
\title{Get Latest Data Freeze and Version}
\usage{
cdm_getLatestDataFreezeAndVersion(connection)
}
\arguments{
\item{connection}{BigQuery connection object}
}
\value{
Character string with the latest data freeze and version (e.g., "r13_v3")
}
\description{
Get Latest Data Freeze and Version
}
