% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection_handler.R
\name{create_fg_connection_handler_FromList}
\alias{create_fg_connection_handler_FromList}
\title{createfg_connection_handlerFromList}
\usage{
create_fg_connection_handler_FromList(config)
}
\arguments{
\item{config}{A list containing configuration settings for the fg_connection_handler.
\itemize{
\item databaseName: The name of the database.
\item connection: A list of connection details settings.
\item cdm: A list of CDM database schema settings.
\item cohortTable: The name of the cohort table.
}}
}
\value{
A fg_connection_handler object.
}
\description{
A function to create a fg_connection_handler object from a list of configuration settings.
}
