% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectAndRunGWASAnalysis.R
\name{createSandboxAPIConnection}
\alias{createSandboxAPIConnection}
\title{FUNCTION_TITLE}
\usage{
createSandboxAPIConnection(base_url, token)
}
\arguments{
\item{base_url}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
