% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{.assertDataFreeze}
\alias{.assertDataFreeze}
\title{Assert Data Freeze}
\usage{
.assertDataFreeze(connection, dataFreeze)
}
\arguments{
\item{connection}{BigQuery connection object}

\item{dataFreeze}{Data freeze identifier to validate}
}
\value{
NULL (called for side effects)
}
\description{
Assert Data Freeze
}
\keyword{internal}
