% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_CDMConnector.R
\name{.assertDataFreezeVersion}
\alias{.assertDataFreezeVersion}
\title{Assert Data Freeze Version}
\usage{
.assertDataFreezeVersion(connection, cdmDataFreezeVersion)
}
\arguments{
\item{connection}{BigQuery connection object}

\item{cdmDataFreezeVersion}{Data freeze version to validate}
}
\value{
NULL (called for side effects)
}
\description{
Assert Data Freeze Version
}
\keyword{internal}
