% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection.R
\name{.assertEnvironment}
\alias{.assertEnvironment}
\title{Assert Environment}
\usage{
.assertEnvironment(environment)
}
\arguments{
\item{environment}{Environment identifier to validate}
}
\value{
NULL (called for side effects)
}
\description{
Assert Environment
}
\keyword{internal}
