% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{.lastNumberSuffix}
\alias{.lastNumberSuffix}
\title{Get Last Number Suffix}
\usage{
.lastNumberSuffix(stringVector, prefix)
}
\arguments{
\item{stringVector}{Vector of strings with number suffixes}

\item{prefix}{Prefix before the number (either "r" or "v")}
}
\value{
Character string with the highest number suffix
}
\description{
Get Last Number Suffix
}
\keyword{internal}
