% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{.lastTableVersion}
\alias{.lastTableVersion}
\title{Get Last Table Version}
\usage{
.lastTableVersion(connection, full_path)
}
\arguments{
\item{connection}{BigQuery connection object}

\item{full_path}{Full path template for the table}
}
\value{
Character string with the latest version (e.g., "v3")
}
\description{
Get Last Table Version
}
\keyword{internal}
