% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fg_append_specialtiy_info_to_service_sector_data.R
\name{fg_append_specialtiy_info_to_service_sector_data_sql}
\alias{fg_append_specialtiy_info_to_service_sector_data_sql}
\title{fg_append_specialtiy_info_to_service_sector_data_sql}
\usage{
fg_append_specialtiy_info_to_service_sector_data_sql(
  service_sector_data_table,
  fg_codes_info_table,
  new_colums_sufix = ""
)
}
\arguments{
\item{fg_codes_info_table}{full path to the table with the codes info}

\item{new_colums_sufix}{string indicating a prefix to add to the appended columns, default="".}

\item{longitudinal_data_table}{full path to the table in longitudinal_data format}
}
\value{
sql script ready to be ran
}
\description{
fg_append_specialtiy_info_to_service_sector_data_sql
}
