% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_append_concept_info_data.R
\name{fg_bq_append_concept_info_data}
\alias{fg_bq_append_concept_info_data}
\title{fg_bq_append_concept_info_data}
\usage{
fg_bq_append_concept_info_data(bq_project_id, bq_table, omop_schema, ...)
}
\arguments{
\item{bq_project_id}{string with the bigquery project id}

\item{bq_table}{an object of type <bq_table> with a table in longitudinal_data format}

\item{omop_schema}{string with the schema where the omop tables are stored}

\item{...}{see \code{fg_append_concept_info_data_sql} for the mapping options}
}
\value{
bq_table with added columns
}
\description{
Wrap around fg_bq_append_concept_info_data_sql to work with bigrquery package
}
