% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fg_append_service_sector_info_to_service_sector_data.R
\name{fg_bq_append_service_sector_info_to_service_sector_data_sql}
\alias{fg_bq_append_service_sector_info_to_service_sector_data_sql}
\title{fg_bq_append_code_info_to_longitudinal_data}
\usage{
fg_bq_append_service_sector_info_to_service_sector_data_sql(
  bq_project_id,
  bq_table,
  fg_codes_info_table,
  ...
)
}
\arguments{
\item{bq_project_id}{string with the bigquery project id}

\item{bq_table}{an object of type <bq_table> with a table in longitudinal_data format}

\item{fg_codes_info_table}{string with the full path (project.schema.table) to the bq table with the fg_codes_info}

\item{...}{see \link{fg_append_code_info_to_longitudinal_data_sql} for the mapping options}
}
\value{
bq_table with added columns
}
\description{
Wrap around fg_append_code_info_to_longitudinal_data_sql to work with bigrquery package
}
