% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{fg_bq_tables}
\alias{fg_bq_tables}
\title{FinnGen BigQuery Tables Handler}
\description{
R6 class for handling BigQuery tables information including environment and data freeze
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{connection}}{BigQuery connection object (read-only).}

\item{\code{environment}}{Environment identifier (e.g., "build", "prod") (read-only).}

\item{\code{dataFreeze}}{Data freeze identifier (e.g., "r13", "dev") (read-only).}

\item{\code{tables}}{Tibble containing table paths information (read-only).}

\item{\code{tbl}}{List of dplyr table objects (read-only).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fg_bq_tables-new}{\code{fg_bq_tables$new()}}
\item \href{#method-fg_bq_tables-clone}{\code{fg_bq_tables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_bq_tables-new"></a>}}
\if{latex}{\out{\hypertarget{method-fg_bq_tables-new}{}}}
\subsection{Method \code{new()}}{
Creates a new fg_bq_tables object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_bq_tables$new(environment, dataFreeze = NULL, tablesPathsTibble = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{environment}}{Environment identifier (e.g., "build", "prod")}

\item{\code{dataFreeze}}{(Optional) Data freeze identifier (default is NULL)}

\item{\code{tablesPathsTibble}}{(Optional) Tibble containing table paths (default is NULL)}

\item{\code{connection}}{BigQuery connection object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_bq_tables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fg_bq_tables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_bq_tables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
