% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_CDMConnector.R
\name{fg_CDMConnector}
\alias{fg_CDMConnector}
\title{Connect to FinnGen OMOP CDM}
\usage{
fg_CDMConnector(environment = NULL, cdmDataFreezeVersion = NULL, ...)
}
\arguments{
\item{environment}{Environment identifier (e.g., "build", "review", or "sandbox-XX")}

\item{cdmDataFreezeVersion}{Data freeze version (e.g., "r13_v3", "dev"). If NULL, latest version is used.}

\item{...}{Additional arguments passed to CDMConnector::cdmFromCon}
}
\value{
A CDM reference object
}
\description{
Connect to FinnGen OMOP CDM
}
