% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection_handler.R
\name{fg_connection_handler}
\alias{fg_connection_handler}
\title{fg_connection_handler}
\description{
Class for handling database connection and schema information for a CDM database
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{databaseName}}{A text id for the database the it connects to (read-only).}

\item{\code{connectionHandler}}{ConnectionHandler object for managing the database connection (read-only).}

\item{\code{sandboxToolsSchema}}{Name of the vocabulary database schema (read-only).}

\item{\code{medicalCodesSchema}}{Name of the CDM database schema (read-only).}

\item{\code{connectionStatusLog}}{Log tibble object for storing connection status information (read-only).}

\item{\code{getTblmedicalCodesSchema}}{List of functions that create dbplyr table for the vocabulary tables (read-only).}

\item{\code{getTblsandboxToolsSchema}}{List of functions that create dbplyr table for the CDM tables (read-only).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fg_connection_handler-new}{\code{fg_connection_handler$new()}}
\item \href{#method-fg_connection_handler-finalize}{\code{fg_connection_handler$finalize()}}
\item \href{#method-fg_connection_handler-loadConnection}{\code{fg_connection_handler$loadConnection()}}
\item \href{#method-fg_connection_handler-clone}{\code{fg_connection_handler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_connection_handler-new"></a>}}
\if{latex}{\out{\hypertarget{method-fg_connection_handler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_connection_handler$new(
  databaseName,
  connectionHandler,
  medicalCodesSchema,
  sandboxToolsSchema
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{A text id for the database the it connects to}

\item{\code{connectionHandler}}{A ConnectionHandler object}

\item{\code{medicalCodesSchema}}{Name of the CDM database schema}

\item{\code{sandboxToolsSchema}}{(Optional) Name of the vocabulary database schema (default is medicalCodesSchema)
Finalize method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_connection_handler-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-fg_connection_handler-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes the connection if active.
Reload connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_connection_handler$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_connection_handler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-fg_connection_handler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Updates the connection status by checking the database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_connection_handler$loadConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fg_connection_handler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fg_connection_handler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fg_connection_handler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
