% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_append_code_info_to_longitudinal_data.R
\name{fg_dbplyr_append_code_info_to_longitudinal_data}
\alias{fg_dbplyr_append_code_info_to_longitudinal_data}
\title{fg_dbplyr_append_code_info_to_longitudinal_data}
\usage{
fg_dbplyr_append_code_info_to_longitudinal_data(
  dbplyr_table,
  fg_bq_tables,
  ...
)
}
\arguments{
\item{dbplyr_table}{an object of type \if{html}{\out{<tbl>}} representing a table in longitudinal_data format}

\item{fg_bq_tables}{an object of type <fg_bq_tables> with the connections to the bigquery tables. The function will use the connection to the bigquery project and the path to fg_codes_info_table to run the query generated by \code{fg_append_code_info_to_longitudinal_data_sql}}

\item{...}{see \url{fg_append_code_info_to_longitudinal_data_sql} for the mapping options}
}
\value{
\if{html}{\out{<tbl>}} with added columns
}
\description{
Wrap around fg_append_code_info_to_longitudinal_data_sql to work with dbplyr package
}
