% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_append_concept_info_data.R
\name{fg_dbplyr_append_concept_info_data}
\alias{fg_dbplyr_append_concept_info_data}
\title{fg_dbplyr_append_concept_info_data}
\usage{
fg_dbplyr_append_concept_info_data(dbplyr_table, fg_bq_tables, ...)
}
\arguments{
\item{dbplyr_table}{a tbl object with the data}

\item{fg_bq_tables}{an object of type fg_bq_tables with the bigquery tables}

\item{...}{see \code{fg_append_concept_info_data_sql} for the mapping options}
}
\value{
tbl with added columns
}
\description{
fg_dbplyr_append_concept_info_data
}
