% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_DatabaseConnector.R
\name{fg_getDatabaseConnector}
\alias{fg_getDatabaseConnector}
\title{Get Database Connector}
\usage{
fg_getDatabaseConnector(environment)
}
\arguments{
\item{environment}{A character string specifying the database environment.
Can be "build" for development/build environment, "preview" for preview environment,
or a sandbox identifier in the format "sandbox-N" where N is the sandbox number.}
}
\value{
A database connector object configured with appropriate connection parameters.
}
\description{
Establishes and returns a database connector object for accessing FinnGen database resources.
}
\details{
This function initializes a database connection using DatabaseConnector package utilities.
The connector can be used to query and interact with FinnGen database tables.
}
