% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{fg_getLatestDataFreeze}
\alias{fg_getLatestDataFreeze}
\title{Get Latest Data Freeze}
\usage{
fg_getLatestDataFreeze(connection)
}
\arguments{
\item{connection}{BigQuery connection object}
}
\value{
Character string with the latest data freeze (e.g., "r13")
}
\description{
Get Latest Data Freeze
}
