% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_get_cdm_config.R
\name{get_cdm_config}
\alias{get_cdm_config}
\title{Get CDM Configuration}
\usage{
get_cdm_config(
  environment,
  dataFreezeNumber,
  cohortTableName = "test_cohort_table",
  atlasDevelopment_gckey =
    "/Users/javier/keys/atlas-development-270609-410deaacc58b.json",
  atlasDevelopment_pathToDriver = "/Users/javier/.config/hades/bigquery",
  asYalm = FALSE
)
}
\arguments{
\item{environment}{Environment must be 'sandbox-' followed by the sandbox number (Alternatively, it can be 'atlasDevelopment' for testing outside sandbox).}

\item{dataFreezeNumber}{The data freeze number to retrieve configuration settings for.}

\item{cohortTableName}{If a cohort table is needed, it can be set here, (default is "test_cohort_table").}

\item{atlasDevelopment_gckey}{If environment = 'atlasDevelopment', the path to the Google Cloud key file.}

\item{atlasDevelopment_pathToDriver}{If environment = 'atlasDevelopment', the path to the BigQuery driver.}

\item{asYalm}{Whether to return the configuration settings as a YAML string (default is FALSE).}
}
\value{
A list of configuration settings for the specified environment and data freeze number.
}
\description{
This function retrieves the configuration settings to work with the OMOP-CDM database for a specified environment and data freeze number.
}
\examples{
get_cdm_config("sandbox-6", 13)

}
