% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTibble.R
\name{LogTibble}
\alias{LogTibble}
\title{LogTibble}
\description{
Class for managing log messages as a tibble.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{logTibble}}{get as tibble}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogTibble-new}{\code{LogTibble$new()}}
\item \href{#method-LogTibble-addLog}{\code{LogTibble$addLog()}}
\item \href{#method-LogTibble-INFO}{\code{LogTibble$INFO()}}
\item \href{#method-LogTibble-WARNING}{\code{LogTibble$WARNING()}}
\item \href{#method-LogTibble-ERROR}{\code{LogTibble$ERROR()}}
\item \href{#method-LogTibble-SUCCESS}{\code{LogTibble$SUCCESS()}}
\item \href{#method-LogTibble-print}{\code{LogTibble$print()}}
\item \href{#method-LogTibble-clone}{\code{LogTibble$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new LogTibble object.
addLog
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-addLog"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-addLog}{}}}
\subsection{Method \code{addLog()}}{
Adds a log message to the log tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$addLog(type, step, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of log message ("INFO", "WARNING", "ERROR", "SUCCESS")}

\item{\code{step}}{Step or description associated with the log message}

\item{\code{message}}{Log message content}

\item{\code{...}}{Additional parameters for message formatting
INFO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-INFO"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-INFO}{}}}
\subsection{Method \code{INFO()}}{
Adds an informational log message to the log tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$INFO(step, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Step or description associated with the log message}

\item{\code{message}}{Log message content}

\item{\code{...}}{Additional parameters for message formatting
WARNING}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-WARNING"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-WARNING}{}}}
\subsection{Method \code{WARNING()}}{
Adds a warning log message to the log tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$WARNING(step, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Step or description associated with the log message}

\item{\code{message}}{Log message content}

\item{\code{...}}{Additional parameters for message formatting
ERROR}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-ERROR"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-ERROR}{}}}
\subsection{Method \code{ERROR()}}{
Adds an error log message to the log tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$ERROR(step, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Step or description associated with the log message}

\item{\code{message}}{Log message content}

\item{\code{...}}{Additional parameters for message formatting
SUCCESS}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-SUCCESS"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-SUCCESS}{}}}
\subsection{Method \code{SUCCESS()}}{
Adds an error log message to the log tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$SUCCESS(step, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Step or description associated with the log message}

\item{\code{message}}{Log message content}

\item{\code{...}}{Additional parameters for message formatting
print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-print"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-print}{}}}
\subsection{Method \code{print()}}{
prints log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogTibble-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogTibble-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogTibble$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
