% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras.R
\name{ResultModelManager_createConnectionHandler}
\alias{ResultModelManager_createConnectionHandler}
\title{createConnectionHandler}
\usage{
ResultModelManager_createConnectionHandler(
  connectionDetailsSettings,
  tempEmulationSchema = NULL,
  useBigrqueryUpload = NULL,
  usePooledConnection = FALSE,
  ...
)
}
\arguments{
\item{connectionDetailsSettings}{A list of connection details settings to pass directly to DatabaseConnector::createConnectionDetails.}

\item{tempEmulationSchema}{The temporary emulation schema (optional).}

\item{useBigrqueryUpload}{Logical indicating whether to use fast table upload for bigquery (default is FALSE)}

\item{usePooledConnection}{Logical indicating whether to use a pooled connection (default is FALSE).}

\item{...}{Additional arguments to be passed to the connection handler object.}
}
\value{
A connection handler object.
}
\description{
Creates a connection handler object based on the provided connection details settings.
}
