% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras.R
\name{tmp_dplyr_copy_to}
\alias{tmp_dplyr_copy_to}
\title{tmp fix for dplyr::copy_to}
\usage{
tmp_dplyr_copy_to(
  dest,
  df,
  name = deparse(substitute(df)),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{dest}{The destination database connection or table name.}

\item{df}{The data frame to be copied.}

\item{name}{The name of the destination table (default is the name of the data frame).}

\item{overwrite}{Logical value indicating whether to overwrite an existing table (default is FALSE).}

\item{...}{pass parameters to dplyr::copy_to}
}
\value{
The new table created or the updated table.
}
\description{
dplyr::copy_to is very slow to upload tables to BQ.
This is function calls dplyr::copy_to except if option "useBigrqueryUpload" option is set to TRUE
In that case it uses package bigrquery to upload the table
}
