% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras.R
\name{tmp_inDatabaseSchema}
\alias{tmp_inDatabaseSchema}
\title{tmp fix for DatabaseConnector::inDatabaseSchema}
\usage{
tmp_inDatabaseSchema(databaseSchema, table)
}
\arguments{
\item{databaseSchema}{The name of the database schema.}

\item{table}{The name of the table.}
}
\value{
The fully qualified table name with the database schema.
}
\description{
till this is fixed https://github.com/OHDSI/DatabaseConnector/issues/236
}
