% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection.R
\name{fg_connection}
\alias{fg_connection}
\title{Create FinnGen BigQuery Connection}
\usage{
fg_connection(enviroment)
}
\arguments{
\item{enviroment}{Environment identifier (e.g., "build", "review", or "sandbox-XX")}
}
\value{
A BigQuery connection object
}
\description{
Create FinnGen BigQuery Connection
}
