% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_append_concept_info_data.R
\name{fg_dbplyr_append_concept_info_data}
\alias{fg_dbplyr_append_concept_info_data}
\title{fg_dbplyr_append_concept_info_data}
\usage{
fg_dbplyr_append_concept_info_data(dbplyr_table, omop_schema, ...)
}
\arguments{
\item{omop_schema}{string with the schema where the omop tables are stored}

\item{tbl}{a tbl object with the data}
}
\value{
tbl with added columns
}
\description{
fg_dbplyr_append_concept_info_data
}
