% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fg_append_visit_type_info_to_service_sector_data.R
\name{fg_dbplyr_append_visit_type_info_to_service_sector_data}
\alias{fg_dbplyr_append_visit_type_info_to_service_sector_data}
\title{fg_dbplyr_append_visit_type_info_to_service_sector_data}
\usage{
fg_dbplyr_append_visit_type_info_to_service_sector_data(
  dbplyr_table,
  dbplyr_fg_codes_info_table,
  ...
)
}
\arguments{
\item{dbplyr_table}{an object of type \if{html}{\out{<tbl>}} representing a table in longitudinal_data format}

\item{dbplyr_fg_codes_info_table}{string with the full path (schema.table) to the database table with the fg_codes_info}

\item{...}{see \url{fg_append_visit_type_info_to_service_sector_data_sql} for the mapping options}
}
\value{
\if{html}{\out{<tbl>}} with added columns
}
\description{
Wrap around fg_append_provider_info_to_service_sector_data_sql to work with dbplyr package
}
