% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{fg_getLatestTablePaths}
\alias{fg_getLatestTablePaths}
\title{Get Latest Table Paths}
\usage{
fg_getLatestTablePaths(
  connection,
  dataFreeze,
  skipDataFreezeValidation = FALSE,
  includeCDMTables = FALSE
)
}
\arguments{
\item{connection}{BigQuery connection object}

\item{dataFreeze}{Data freeze identifier}

\item{skipDataFreezeValidation}{Whether to skip data freeze validation}

\item{includeCDMTables}{Whether to include OMOP CDM tables (default is FALSE)}
}
\value{
Tibble with table_id and full_path columns
}
\description{
Get Latest Table Paths
}
