% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bigquery.R
\name{get_fg_bq_tables}
\alias{get_fg_bq_tables}
\title{Get FinnGen BigQuery Tables}
\usage{
get_fg_bq_tables(
  environment,
  dataFreeze = NULL,
  tablesPathsTibble = NULL,
  includeCDMTables = FALSE
)
}
\arguments{
\item{environment}{Environment identifier (e.g., "build", "prod")}

\item{dataFreeze}{(Optional) Data freeze identifier (default is NULL)}

\item{tablesPathsTibble}{(Optional) Tibble containing table paths (default is NULL)}

\item{includeCDMTables}{(Optional) Whether to include OMOP CDM tables (default is FALSE)}
}
\value{
An fg_bq_tables R6 object
}
\description{
Creates a new fg_bq_tables object
}
