% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectAndRunGWASAnalysis.R
\name{runGWASAnalysis}
\alias{runGWASAnalysis}
\title{runGWASAnalysis}
\usage{
runGWASAnalysis(
  connection_sandboxAPI,
  cohorts_settings,
  phenotype_name,
  title = phenotype_name,
  description = phenotype_name,
  cases_description = cohorts_settings$cases_cohort$name,
  controls_description = cohorts_settings$controls_cohort$name,
  notification_email = connection_sandboxAPI$notification_email,
  release = "finngen_R12"
)
}
\arguments{
\item{connection_sandboxAPI}{PARAM_DESCRIPTION}

\item{cohorts_settings}{PARAM_DESCRIPTION}

\item{phenotype_name}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION, Default: phenotype_name}

\item{description}{PARAM_DESCRIPTION, Default: phenotype_name}

\item{cases_description}{PARAM_DESCRIPTION, Default: cohorts_settings$cases_cohort$name}

\item{controls_description}{PARAM_DESCRIPTION, Default: cohorts_settings$controls_cohort$name}

\item{notification_email}{PARAM_DESCRIPTION, Default: connection_sandboxAPI$notification_email}

\item{release}{PARAM_DESCRIPTION, Default: 'Regenie9'}
}
\description{
runGWASAnalysis
}
